fix: revert and remove export of fn
